refactor db

+++ hackathon-service/src/main/java/com/hacof/hackathon/dto/AwardDTO.java
@@ -7,7 +7,7 @@ public class AwardDTO {
     private Long id;
     private String name;
     private String description;
-    private Double prizeAmount;
+    private int prizeAmount;
     private Long hackathonId;
     private String createdBy;
     private String lastUpdatedBy;
